Update to sync changes in push service protocol
The initial HTTP request that does the websocket setup
can now return status 401 (unauthorized) and 429 (too many requests).

The case where 401 is returned replaces the error 4001 that was previously
sent over the established websocket. Now the websocket will not be succesfully
established if client is not authorized.
The reconnect logic has been updated to reflect these changes.